+++ spec/dummy/app/matestack/dummy/pages/products/index.rb
@@ -52,7 +52,7 @@ class Dummy::Pages::Products::Index < ApplicationPage
 
   def collection_card product
     bs_card title: product.name, subtitle: "#{product.price_in_euro} €", class: "h-100" do
-      paragraph product.description, class: "fw-lighter" 
+      paragraph product.description, class: "fw-lighter"
       transition path: edit_dummy_product_path(product), delay: 300 do
         bs_btn outline: true, size: :sm, variant: :primary do
           bs_icon name: 'arrow-right', size: 20