+++ spec/dummy/config/webpack/environment.js
@@ -6,7 +6,9 @@ const webpack = require('webpack');
 const customWebpackConfig = {
   resolve: {
     alias: {
-      'vue': 'vue/dist/vue.esm-bundler'
+      'vue': 'vue/dist/vue.esm-bundler',
+      'matestack-ui-vuejs': 'matestack-ui-vuejs/lib/matestack/ui/vue_js/index.js', // in order not to use the esm package
+      'matestack-ui-bootstrap': 'matestack-ui-bootstrap/lib/matestack/ui/bootstrap/index.js' // in order not to use the esm package
     }
   },
   plugins: [