+++ spec/dummy/db/migrate/20220610134737_add_status_to_orders.rb
@@ -1,0 +1,5 @@
+class AddStatusToOrders < ActiveRecord::Migration[7.0]
+  def change
+    add_column :orders, :status, :integer, default: 0
+  end
+end